wip: refactor strict mode
